Initial validation changes to model files

+++ backend/db/migrations/03-reviews.js
@@ -36,7 +36,7 @@ module.exports = {
         allowNull: false
       },
       stars: {
-        type: Sequelize.INTEGER(5),
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       createdAt: {
